chore(coordinator): persistence package tidy up
