+++ coordinator/ethereum/blob-submitter/src/integrationTest/kotlin/net/consensys/zkevm/ethereum/finalization/BlobAndAggregationFinalizationIntTest.kt
@@ -12,6 +12,10 @@ import linea.domain.BlobSubmittedEvent
 import linea.domain.FinalizationSubmittedEvent
 import linea.persistence.AggregationsRepository
 import linea.persistence.BlobsRepository
+import linea.persistence.conflation.AggregationsRepositoryImpl
+import linea.persistence.conflation.BlobsPostgresDao
+import linea.persistence.conflation.BlobsRepositoryImpl
+import linea.persistence.conflation.PostgresAggregationsDao
 import net.consensys.FakeFixedClock
 import net.consensys.linea.ethereum.gaspricing.FakeGasPriceCapProvider
 import net.consensys.linea.testing.submission.loadBlobsAndAggregations
@@ -21,10 +25,6 @@ import net.consensys.zkevm.ethereum.MakeFileDelegatedContractsManager
 import net.consensys.zkevm.ethereum.coordination.EventDispatcher
 import net.consensys.zkevm.ethereum.submission.BlobSubmissionCoordinator
 import net.consensys.zkevm.ethereum.submission.L1ShnarfBasedAlreadySubmittedBlobsFilter
-import net.consensys.zkevm.persistence.dao.aggregation.AggregationsRepositoryImpl
-import net.consensys.zkevm.persistence.dao.aggregation.PostgresAggregationsDao
-import net.consensys.zkevm.persistence.dao.blob.BlobsPostgresDao
-import net.consensys.zkevm.persistence.dao.blob.BlobsRepositoryImpl
 import net.consensys.zkevm.persistence.db.DbHelper
 import net.consensys.zkevm.persistence.db.test.CleanDbTestSuiteParallel
 import org.assertj.core.api.Assertions.assertThat
