+++ coordinator/persistence/src/main/kotlin/linea/persistence/conflation/BatchesDao.kt
@@ -1,0 +1,14 @@
+package linea.persistence.conflation
+
+import linea.domain.Batch
+import tech.pegasys.teku.infrastructure.async.SafeFuture
+
+interface BatchesDao {
+  fun saveNewBatch(batch: Batch): SafeFuture<Unit>
+
+  fun findHighestConsecutiveEndBlockNumberFromBlockNumber(startingBlockNumberInclusive: Long): SafeFuture<Long?>
+
+  fun deleteBatchesUpToEndBlockNumber(endBlockNumberInclusive: Long): SafeFuture<Int>
+
+  fun deleteBatchesAfterBlockNumber(startingBlockNumberInclusive: Long): SafeFuture<Int>
+}
