+++ coordinator/persistence/src/main/kotlin/linea/persistence/conflation/BlobsDao.kt
@@ -1,0 +1,22 @@
+package linea.persistence.conflation
+
+import linea.domain.BlobRecord
+import tech.pegasys.teku.infrastructure.async.SafeFuture
+import kotlin.time.Instant
+
+interface BlobsDao {
+  fun saveNewBlob(blobRecord: BlobRecord): SafeFuture<Unit>
+
+  fun getConsecutiveBlobsFromBlockNumber(
+    startingBlockNumberInclusive: ULong,
+    endBlockCreatedBefore: Instant,
+  ): SafeFuture<List<BlobRecord>>
+
+  fun findBlobByStartBlockNumber(startBlockNumber: ULong): SafeFuture<BlobRecord?>
+
+  fun findBlobByEndBlockNumber(endBlockNumber: ULong): SafeFuture<BlobRecord?>
+
+  fun deleteBlobsUpToEndBlockNumber(endBlockNumberInclusive: ULong): SafeFuture<Int>
+
+  fun deleteBlobsAfterBlockNumber(startingBlockNumberInclusive: ULong): SafeFuture<Int>
+}
