--- coordinator/persistence/src/main/kotlin/net/consensys/zkevm/persistence/dao/batch/persistence/BatchesDao.kt
@@ -1,14 +1,0 @@
-package net.consensys.zkevm.persistence.dao.batch.persistence
-
-import linea.domain.Batch
-import tech.pegasys.teku.infrastructure.async.SafeFuture
-
-interface BatchesDao {
-  fun saveNewBatch(batch: Batch): SafeFuture<Unit>
-
-  fun findHighestConsecutiveEndBlockNumberFromBlockNumber(startingBlockNumberInclusive: Long): SafeFuture<Long?>
-
-  fun deleteBatchesUpToEndBlockNumber(endBlockNumberInclusive: Long): SafeFuture<Int>
-
-  fun deleteBatchesAfterBlockNumber(startingBlockNumberInclusive: Long): SafeFuture<Int>
-}
